requirements.txt
pandas==1.3.3
numpy==1.21.2
azure-functions==1.13.2